// src/types/Permission.ts
export enum Permission {
  ViewUsers = 1,
  ManageUsers = 2,
  ViewAccessGroups = 3,
  ManageAccessGroups = 4,
  ViewCategories = 5,
  ManageCategories = 6,
  ViewPaymentMethods = 7,
  ManagePaymentMethods = 8,
  ViewMerchandises = 9,
  ManageMerchandises = 10,
  ViewSales = 11,
  ManageSales = 12,
};